drop dependency on Crypt::OpenSSL::CA
We were only using this in the test suite to generate the X509
certificate that the test suite was using.  The problem is,
Crypt::OpenSSL::CA has not been updated for OpenSSL v1.1, and as a
result Crypt::OpenSSL::CA fails to install on a lot of systems.

Instead we are now just using the openssl command via
IPC::System::Simple to generate a self signed certificate

Fixes #1