eliminate in-memory keystore, and use wrapped key handles
We no longer use an in-memory keystore to keep track of generated
handles.  Instead, key handles that are generated by register() are
"wrapped prviate keys" as allowed in the FIDO/U2F specs.  The way this
is implemented is the raw private key that was generated for the
register() request is encrypted using the Tester private key.

This allows DEVICE_INELIGIBLE to be handled by the following check:
    - register()
        if any of the passed in registered key handles successfully
        decrypt, return DEVICE_INELIGIBLE.

    - sign()
        If none of the passed in key handles can be decrypted, then
        return DEVICE_INELIGIBLE.

This also adds a method:

    is_known_handle($handle): bool

        returne true if the given handle was generated by this tester.
        $handle is a Base64 URL encoded key handle.  This method will
        return true if the handle can be decrypted using the testers
        private key.