move keystore management into its own class / role
This moves "keystore" management into its own class and allows for
different keystore strategies to be implemented as keystore classes in
the future.  The default keystore is
Authen::U2F::Tester::Keystore::Wrapped, but you can now implement your
own keystore by implementing the Authen::U2F::Tester::Role::Keystore
role and passing in the keystore to the Tester constructor.  For
example:

 my $keystore = My::Authen::U2F::Tester::Keystore->new(...);

 my $tester = Authen::U2F::Tester->new(
    certificate => ...,
    key         => ...,
    keystore    => $keystore);

If no keystore is given in the args to new(), then the default keystore
will be used.