Turn on "User Presence" bit in sign() responses
This makes Authen::U2F::Tester work with Yubico's libu2f-server library,
which has a hard-coded requirement that the user presence byte is "1".
Previously we sent 0x00, which indicates we did not confirm the presence
of the user, but it seems we should send 0x01 in order to be compatible
with more devices / implementations.

Fixes #2